Removing multi-server feature and updating LibertyExtension (#373)
* Removing multi-server feature, adding CompileJSPExtension, updating LibertyExtension

* Adding missing extension and updating docs
