+++ src/main/groovy/io/openliberty/tools/gradle/extensions/LibertyExtension.groovy
@@ -23,90 +23,35 @@ class LibertyExtension {
     String installDir
     String outputDir
     String userDir
-    String serverName = "defaultServer"
+
     String baseDir
     String cacheDir
-
-    String appsDirectory = "apps"
-    boolean stripVersion = false
-    boolean looseApplication = true
-
-    File configDirectory
-    File serverXmlFile
-    File bootstrapPropertiesFile
-    File jvmOptionsFile
-    File serverEnvFile
-
-    boolean clean = false
-    String timeout
-    String template
-    boolean noPassword = false
-    boolean embedded = false
-
-    int verifyAppStartTimeout = 0
-
-    String jspVersion
-    int jspCompileTimeout = 40
 
     // For overriding the group, name or version of the libertyRuntime installed from Maven Central repository.
     // Default is group 'io.openliberty', name 'openliberty-kernel' and version '[19.0.0.9,)' which gets the latest version.
     Properties runtime = new Properties()
 
-    def numberOfClosures = 0
+    CompileJSPExtension jsp = new CompileJSPExtension()
 
-    FeatureExtension features = new FeatureExtension()
-    UninstallFeatureExtension uninstallfeatures = new UninstallFeatureExtension()
     InstallExtension install = new InstallExtension()
-    CleanExtension cleanDir = new CleanExtension()
     SpringBootExtension thin = new SpringBootExtension()
 
-    PackageExtension packageLiberty = new PackageExtension()
-    DumpExtension dumpLiberty = new DumpExtension()
-    DumpExtension javaDumpLiberty = new DumpExtension()
+    ServerExtension server = server = new ServerExtension()
 
-    NamedDomainObjectContainer<ServerExtension> servers
-
-    ServerExtension server
-
-    def uninstallfeatures(Closure closure) {
-        ConfigureUtil.configure(closure, uninstallfeatures)
-    }
-
-    def features(Closure closure) {
-        ConfigureUtil.configure(closure, features)
-    }
-
-    def install(Closure closure) {
-        ConfigureUtil.configure(closure, install)
-    }
-
-    def packageLiberty(Closure closure) {
-        ConfigureUtil.configure(closure, packageLiberty)
-    }
-
-    def dumpLiberty(Closure closure) {
-        ConfigureUtil.configure(closure, dumpLiberty)
-    }
-
-    def javaDumpLiberty(Closure closure) {
-        ConfigureUtil.configure(closure, javaDumpLiberty)
-    }
-
-    def cleanDir(Closure closure) {
-        ConfigureUtil.configure(closure, cleanDir)
+    def jsp(Closure closure) {
+        ConfigureUtil.configure(closure, jsp)
     }
 
     def thin(Closure closure) {
        ConfigureUtil.configure(closure, thin)
     }
 
-    def server(Closure closure){
-        server = new ServerExtension()
-        ConfigureUtil.configure(closure, server)
+    def install(Closure closure) {
+        ConfigureUtil.configure(closure, install)
     }
 
-    def servers(Closure configurationClosure){
-        servers.configure(configurationClosure)
+    def server(Closure closure){
+        ConfigureUtil.configure(closure, server)
     }
 
 }
