+++ src/main/groovy/io/openliberty/tools/gradle/tasks/CompileJSPTask.groovy
@@ -62,7 +62,7 @@ class CompileJSPTask extends AbstractServerTask {
         compileJsp.setInstallDir(getInstallDir(project))
         compileJsp.setTempdir(project.buildDir)
         compileJsp.setDestdir(new File(project.buildDir.getAbsolutePath()+"/classes/java"))
-        compileJsp.setTimeout(project.liberty.jspCompileTimeout)
+        compileJsp.setTimeout(project.liberty.jsp.jspCompileTimeout)
         // don't delete temporary server dir
         compileJsp.setCleanup(false)
         compileJsp.setProject(ant)
@@ -86,8 +86,8 @@ class CompileJSPTask extends AbstractServerTask {
         logger.debug("Classpath: " + classpathStr)
         compileJsp.setClasspath(classpathStr)
 
-        if (project.liberty.jspVersion != null) {
-            compileJsp.setJspVersion(project.liberty.jspVersion)
+        if (project.liberty.jsp.jspVersion != null) {
+            compileJsp.setJspVersion(project.liberty.jsp.jspVersion)
         }
 
         compileJsp.init()
