+++ src/test/resources/liberty-package-test/liberty-package-archiveZip.gradle
@@ -27,7 +27,6 @@ dependencies {
 liberty {
     server{
         name = 'LibertyProjectServer'
-        name = serverName
         packageLiberty {
             packageDirectory = "${project.projectDir}"
             packageName = 'testPackage'
