--- src/test/resources/liberty-test/oldStyleTest.gradle
@@ -1,43 +1,0 @@
-apply plugin: 'liberty'
-
-buildscript {
-    repositories {
-        mavenLocal()
-        mavenCentral()
-        maven {
-            name = 'Sonatype Nexus Snapshots'
-            url = 'https://oss.sonatype.org/content/repositories/snapshots/'
-        }
-    }
-    dependencies {
-        classpath fileTree(dir: '../libs', include: '*.jar', exclude: ['*-groovydoc.jar', '*-sources.jar'])
-        classpath "io.openliberty.tools:liberty-gradle-plugin:$lgpVersion"
-    }
-}
-
-repositories {
-    mavenLocal()
-    mavenCentral()
-}
-
-dependencies {
-    libertyRuntime group: runtimeGroup, name: runtimeArtifactId, version: runtimeVersion
-}
-
-liberty {
-    serverName ='LibertyProjectServer'
-
-    packageLiberty {
-        include = 'minify'
-        os='Linux'
-    }
-
-    features {
-        name= ['mongodb-2.0']
-        acceptLicense = true
-    }
-
-    uninstallfeatures {
-        name= ['mongodb-2.0']
-    }
-}
